Added missing casts for constructing 32-bit values from bytes
A lack of these casts can cause data corruption on 16-bit microcontrollers.

+++ tmc/ic/TMC2660/TMC2660.c
@@ -81,7 +81,7 @@ void readWrite(uint8_t icID, uint32_t datagram)
     // Send 24 bytes of data and receive reply
     tmc2660_readWriteSPI(icID, &data[0], sizeof(data));
 
-    reply = (data[0] << 16 | data[1] << 8 | data[2]) >> 4;
+    reply = (((uint32_t) data[0] << 16) | ((uint32_t) data[1] << 8) | data[2]) >> 4;
 
     // write value to response shadow register
     tmc2660_shadowRegister[icID][rdsel] = reply;
